Code cleanup & Fix to proxy in porduction
related: https://github.com/Teranode/nuxt-module-alternatives/issues/5
