+++ @nuxtjs-alt/proxy/readme.md
@@ -4,9 +4,7 @@ This serves as an alternative for @nuxtjs-alt/proxy. Please note that his is for
 
 **Other Information**
 
-This module works a litte bit differently as the `addServerMiddleware` function does not want to accept non-files, so instead this module ~~creates a folder in your `srcDir` (project folder) called `proxy-middleware` and for each proxy entry in your nuxt options creates a file called `proxy-<index>.ts` with `index` being the placement within that proxy object. This is as much as i could do in terms of making it work with vite and nuxt 3, if you want to make any changes to it feel free to submit an issue.~~
-
-This will now create files in it's own self contained directory in the runtime folder in the module.
+This module works a litte bit differently as the `addServerMiddleware` function does not want to accept non-files, so instead this module creates a folder in your `srcDir` (project folder) inside `server` called `proxy` and for each proxy entry in your nuxt options creates a file called `proxy-<index>.ts` with `index` being the placement within that proxy object. The files are auto-generated. This is as much as i could do in terms of making it work with vite and nuxt 3, if you want to make any changes to it feel free to submit an issue. 
 
 **Depenencies Needed:**
 - http-proxy-middleware
